chore: update dependencies, upgrade Gradle to 9.5.0, simplify null-safe data handling, and add Foojay toolchain support.

+++ app/src/main/java/com/sameerasw/essentials/complication/MainComplicationService.kt
@@ -18,7 +18,7 @@ class MainComplicationService : SuspendingComplicationDataSourceService() {
         return when (type) {
             ComplicationType.SHORT_TEXT -> createComplicationData("#FFFFFFFF", "Theme Color")
             ComplicationType.RANGED_VALUE -> createRangedComplicationData(
-                0xFFFFFFFF.toLong(),
+                0xFFFFFFFF,
                 "Theme Color"
             )
